Store `MethodInfo` when creating an `ActivityDefinition`
In order to access other attributes for a given activity, the MethodInfo used to create the activity is needed. There are only certain situations where a MethodInfo is used so the new property is added but made nullable.

New tests are added to verify each public Create method properly sets or does not set the MethodInfo.